Fixed circulos. Agregado por fin la conversion de unidades al atributo del drawer. Ahora dibujamos circulos. Si la conversion de unidades para X y para Y son correspondientes, queda un circulo.

+++ tallerDeProgramacionI-SnowBros/headers/Modelo/Objetos/Figuras/Rectangulo.h
@@ -5,8 +5,8 @@
 
 class Rectangulo : public Figura {
 private:
-	int ancho;
-	int alto;
+	float ancho;
+	float alto;
 public:
 	Rectangulo(JsonParser *parser, int index, b2World* world);
